requirements.txt
amino.py
aiogram
json_minify